Add files via upload
In Today’s world Sensitive data is increasingly used in communication over the internet. Thus Security of data is the biggest concern of internet users. Best solution is use of some cryptography algorithm which encrypts data in some cipher and transfers it over the internet and again decrypted to original data. The field of cryptography deals with the procedure for conveying information securely. The goal is to allow the intended recipients of a message to receive the message properly while interrupt eaves- droppers from understanding the message. Cryptography includes a set of techniques for scrambling or disguising data so that it is available only to someone who can restore the data to its original form. In current computer systems, cryptography provides a strong, economical basis for keeping data classified and for verifying data indignity. While our conventional cryptography methods, such for AES (encryption) and RSA (signing), work well on systems which have reasonable processing power and memory capabilities, these do not scale well into a world with embedded systems and sensor networks. Thus, lightweight cryptography methods are proposed to overcome many of the problems of conventional cryptography. This paper sets out to contribute to the general body of knowledge in the area of classical cryptography by developing a new hybrid way of encryption of plaintext. The cryptosystem performs its encryption by encrypting the plaintext using Vigenere Cipher and further using the ciphertext to encrypt the plaintext again using Polybius Cipher.